One tiny error in readme
The whitespace ` ` character in headings like `#### ` seems to sometimes be invalid. Not sure if this is an encoding issues or what. It might be worth investigating to see if Github Flavored Markdown will support these cases. Definitely odd!

P.S. this might be a contender for the _smallest PR ever made_.